remodel of extension
it seems that storing runtime data in the background page is bad because
the background page becomes inactive (chrome does this to save ram/cpu)
and all the runtime data gets deleted. now I just use the background
page to show the icon and notify the content script when button is
pressed. on the other hand, dynamic webpages (like youtube internal
load) stopped working.
anyway, this should be better.